+++ AL-Game/src/com/aionemu/gameserver/utils/SafeMath.java
@@ -14,6 +14,7 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.aionemu.gameserver.utils;
 
 /**
@@ -31,7 +32,7 @@ public class SafeMath {
 
 	public static long addSafe(long source, long value) throws OverfowException {
 		if ((source > 0 && value > Long.MAX_VALUE - source) || (source < 0 && value < Long.MIN_VALUE - source)) {
-			throw new OverfowException(source + " + " + value + " = " + (source + value));
+			throw new OverfowException(source + " + " + value + " = " + ((long) source + (long) value));
 		}
 		return source + value;
 	}
@@ -51,42 +52,34 @@ public class SafeMath {
 		if (a > b) {
 			// use symmetry to reduce boundry cases
 			ret = multSafe(b, a);
-		}
-		else {
+		} else {
 			if (a < 0) {
 				if (b < 0) {
 					// check for positive overflow with negative a, negative b
 					if (a >= Long.MAX_VALUE / b) {
 						ret = a * b;
-					}
-					else {
+					} else {
 						throw new OverfowException(msg);
 					}
-				}
-				else if (b > 0) {
+				} else if (b > 0) {
 					// check for negative overflow with negative a, positive b
 					if (Long.MIN_VALUE / b <= a) {
 						ret = a * b;
-					}
-					else {
+					} else {
 						throw new OverfowException(msg);
 
 					}
-				}
-				else {
+				} else {
 					ret = 0;
 				}
-			}
-			else if (a > 0) {
+			} else if (a > 0) {
 				// check for positive overflow with positive a, positive b
 				if (a <= Long.MAX_VALUE / b) {
 					ret = a * b;
-				}
-				else {
+				} else {
 					throw new OverfowException(msg);
 				}
-			}
-			else {
+			} else {
 				ret = 0;
 			}
 		}
