+++ AL-Game/src/com/aionemu/gameserver/utils/audit/AutoBan.java
@@ -14,12 +14,13 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.aionemu.gameserver.utils.audit;
 
 import com.aionemu.gameserver.configs.main.PunishmentConfig;
 import com.aionemu.gameserver.model.gameobjects.player.Player;
-import com.aionemu.gameserver.network.BannedMacManager;
 import com.aionemu.gameserver.network.aion.serverpackets.SM_QUIT_RESPONSE;
+import com.aionemu.gameserver.network.BannedMacManager;
 import com.aionemu.gameserver.network.loginserver.LoginServer;
 import com.aionemu.gameserver.services.PunishmentService;
 
@@ -37,7 +38,7 @@ public class AutoBan {
 		int playerId = player.getObjectId();
 		int time = PunishmentConfig.PUNISHMENT_TIME;
 		int minInDay = 1440;
-		int dayCount = (int) (Math.floor(time / minInDay));
+		int dayCount = (int) (Math.floor((double) (time / minInDay)));
 
 		switch (PunishmentConfig.PUNISHMENT_TYPE) {
 			case 1:
